+++ src/frontend/sider/Fagsak/Behandling/Sider/Vedtak/VedtakBegrunnelserTabell/endringstidspunkt/useEndringstidspunkt.tsx
@@ -5,7 +5,7 @@ import { useFelt, useSkjema } from '@navikt/familie-skjema';
 import type { Ressurs } from '@navikt/familie-typer';
 import { byggHenterRessurs, hentDataFraRessurs, RessursStatus } from '@navikt/familie-typer';
 
-import { useVedtakStegContext } from '../../../../../../../context/behandlingContext/useVedtakSteg';
+import { useVedtakStegContext } from '../../../../../../../context/behandlingContext/VedtakStegContext';
 import type { IBehandling } from '../../../../../../../typer/behandling';
 import type { IRestOverstyrtEndringstidspunkt } from '../../../../../../../typer/vedtaksperiode';
 import type { IsoDatoString } from '../../../../../../../utils/dato';
